requirements.txt
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
numpy
scipy
scikit-learn
matplotlib
pandas
flasgger
streamlit
tensorflow-cpu
opencv-python-headless
nltk
Pillow
keras
tensorflow_hub
